+++ build.lua
@@ -3,14 +3,21 @@
 
 -- Configuration file of "citeproc" for use with "l3build"
 
-module = "citeproc"
+module = "csl"
 
 docfiledir = "./doc"
 sourcefiledir = "./citeproc"
 testfiledir = "./test"
 
-scriptfiles = {"*.lua", "*.json"}
+installfiles = {"*.sty", "*.lua", "*.json"}
+scriptfiles = {"citeproc*.lua"}
 -- scriptmanfiles = {"citeproc.1"}
-sourcefiles = {"*.lua", "*.json", "*.sty"}
+sourcefiles = {"*.sty", "*.lua", "*.json"}
 -- tagfiles = {}
 -- typesetdemofiles = {}
+
+tdslocations = {
+  "scripts/csl/citeproc/citeproc*.lua",
+  "scripts/csl/*.lua",
+  "tex/latex/csl/citeproc/*.json",
+}